Fix fails on unit tests in Python 3
The unit tests assumed that the ordering of elements in a dict is
deterministic. This is true in Python 2, but not the case in Python 3.

See https://docs.python.org/3/reference/datamodel.html#object.__hash__
for more details.

Specifically, in Python 3 the order of the query parameters in the URL changed
after every run. e.g. "http://www.example.com/index?q1=v1&q2=v2" could
change to http://www.example.com/index?q2=v2&q1=v1". This caused many
unit tests to fail. To fix this, the query parameters are sorted before
URLs are compared.