Use JSON data instead of query paramaters PUT/POST
This is primarily an attempt to make the translation from Python
data types closer to what DO API expects.

In particular I noticed that booleans are inserted into params
using str() so True -> 'True' and False -> 'False'

However the API docs show that boolean values in the query
params should conform to the JSON standard - i.e. 'true'
or 'false'

I believe that the best way to handle this is to just submit
the params data as JSON instead of as query paramaters. In
fact JSON is recommended in the docs -
https://developers.digitalocean.com/#parameters

This approach means that translation is handled for us by
the JSON encoding.
Additionally the requests library can handle this for us
natively - we don't even need to explicitely encode the
data - we just pass it as the json paramater of the relevant
requests method.

This was prompted by a issue I saw where droplets were being
created without private networking even though I had sepecified
private_networking=True in the Droplet.create() call.
When I compared the API call made by the digitalocean module to the docs
I realised that the only difference was in the case of the value -
"True" versus "true"

When I applied this commit the problem went away.

I suspect that a change in the JSON handling in the DO API caused this
to become a problem.